+++ frontend/src/const/time-formats.js
@@ -1,0 +1,1 @@
+export const DD_MMM_YYYY = 'DD MMM YYYY';