+++ frontend/src/utils/validation.js
@@ -1,6 +1,5 @@
-import {validateWebHook} from '../actions/deploy-contract';
-
-const solidityReservedWords = ['address', 'contract', 'function', 'struct', 'uint', 'returns', 'abstract', 'after', 'case', 'catch', 'final', 'in', 'inline', 'interface', 'let', 'match', 'of', 'pure', 'relocatable', 'static', 'switch', 'try', 'type', 'typeof', 'view', 'index', 'storage', 'state', 'variable', 'mapping', 'block', 'coinbase', 'difficulty', 'number', 'block', 'number', 'timestamp', 'msg', 'data', 'gas', 'sender', 'value', 'now', 'gas', 'price', 'origin', 'keccak256', 'ripemd160', 'sha256', 'ecrecover', 'addmod', 'mulmod', 'cryptography', 'this', 'super', 'selfdestruct', 'balance', 'send'];
+import { validateWebHook } from '../actions/deploy-contract';
+import { solidityReservedWords } from '../const/solidity-reserved-words';
 
 const urlErrorMessage = 'Webhook needs to be url with format [protocol]://[url]/[path]';
 
@@ -14,11 +13,22 @@ export const validateScaffoldProperties = values => {
     const scaffoldFieldsErrors = [];
 
     if (field.name) {
-      if (field.name[0].match(/[a-z]/) === null) scaffoldFieldsErrors.push('A property should begin with a lowercase letter');
-      if (field.name.match(/[\s\/\\]/) !== null) scaffoldFieldsErrors.push('A property should not contain a space, / and \\'); // eslint-disable-line
-      if (solidityReservedWords.includes(field.name)) scaffoldFieldsErrors.push(`${field.name} is a reserved word, pick another property name.`);
+      if (field.name[0].match(/[a-z]/) === null)
+        scaffoldFieldsErrors.push('A property should begin with a lowercase letter');
+      /* eslint-disable */
+      if (field.name.match(/[\s\/\\]/) !== null)
+        /* eslint-enable */
+        scaffoldFieldsErrors.push('A property should not contain a space, / and \\');
+      if (solidityReservedWords.includes(field.name))
+        scaffoldFieldsErrors.push(`${field.name} is a reserved word, pick another property name.`);
       if (propertyNames.filter(it => it === field.name).length > 1) scaffoldFieldsErrors.push('Name must be unique');
+    } else {
+      // scaffoldFieldsErrors.push('Name is required');
     }
+
+    // if (!field.type) {
+    // scaffoldFieldsErrors.push('Type is required');
+    // }
 
     const ifNumber = field.type === 'NUMBER';
     const regexTest = field.defaultValue === undefined ? false : digitRegex.test(field.defaultValue);
@@ -34,7 +44,7 @@ export const validateScaffoldProperties = values => {
 
 export const validateAddress = address => {
   let errors = [];
-  if (!address.startsWith('0x')) errors.push('A developer address should beging with 0x');
+  if (!address.startsWith('0x')) errors.push('A developer address should begin with 0x');
   if (address.length !== 42) errors.push('A developer address should be 42 characters long');
   return errors;
 };
@@ -53,13 +63,12 @@ export const warn = values => {
   return warnings;
 };
 
-const isUrl = (str) => {
-  const regexp =  /^((https?|ftp|smtp):\/\/)(www.)?[a-z0-9]+\.[a-z]+(\/[a-zA-Z0-9#]+\/?)*$/;
+const isUrl = str => {
+  const regexp = /^((https?|ftp|smtp):\/\/)(www.)?[a-z0-9]+\.[a-z]+(\/[a-zA-Z0-9#]+\/?)*$/;
   return regexp.test(str);
 };
 
-export const asyncValidate = async (values) => {
-
+export const asyncValidate = async values => {
   if (!values.webHook) {
     return;
   }
@@ -67,7 +76,7 @@ export const asyncValidate = async (values) => {
   try {
     return await validateWebHook(values.webHook);
   } catch (e) {
-    throw {webHook: urlErrorMessage} // eslint-disable-line
+    throw { webHook: urlErrorMessage }; // eslint-disable-line
   }
 };
 
@@ -81,8 +90,8 @@ export const validate = values => {
   if (!values.developerAddress) {
     errors.developerAddress = 'A developer address is required.';
   }
-  if (!values.scaffoldDescription) {
-    errors.scaffoldDescription = 'A scaffold title is required.';
+  if (!values.description) {
+    errors.description = 'A scaffold title is required.';
   }
   if (!values.fiatAmount) {
     errors.fiatAmount = 'Fiat Amount is required.';
@@ -97,9 +106,8 @@ export const validate = values => {
   if (values.properties) {
     const scaffoldFieldErrors = {};
     values.properties.forEach((field, fieldIndex) => {
-      if (!field.property) scaffoldFieldErrors.property = 'required';
-      if (!field.datatype) scaffoldFieldErrors.datatype = 'required';
-      if (!field.defaultValue) scaffoldFieldErrors.defaultValue = 'required';
+      if (!field.name) scaffoldFieldErrors.name = 'required';
+      if (!field.type) scaffoldFieldErrors.type = 'required';
       errors.properties[fieldIndex] = scaffoldFieldErrors;
     });
   }