+++ CHANGELOG.md
@@ -5,8 +5,25 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.3.0] - 2018-06-29
 ### Added
-- Unit tests for services, components, utils
+- Edit scaffold description
+- Developers keys management:
+- Generate API key
+- Show API keys list
+- Add API keys name
+- Edit API keys name
+- Deactivate API key
+- Communication protocol between API and scaffold
+- Scaffold events decoder
+- Backend: events are sent by a webhook
+- Scaffold: unit tests
+- Application unit tests
+
+## Changed
+- Default values are no longer required for additional scaffold properties
+- Application will automatically reconnect to a chain of a connection is lost
 
 ## [2.2.1] - 2018-06-25
 ### Added
@@ -88,6 +105,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - "Withdraw" button is removed
 
 [Unreleased]: https://github.com/OpenFuturePlatform/open-api/compare/master...sprint
+[2.3.0]: https://github.com/OpenFuturePlatform/open-api/compare/v2.2.1...v2.3.0
 [2.2.1]: https://github.com/OpenFuturePlatform/open-api/compare/v2.2.0...v2.2.1
 [2.2.0]: https://github.com/OpenFuturePlatform/open-api/compare/v2.1.0...v2.2.0
 [2.1.0]: https://github.com/OpenFuturePlatform/open-api/compare/v2.0.0...v2.1.0